# swagger-spec.json
{"openapi":"3.0.0","paths":{"/api/whitelist":{"post":{"operationId":"WhitelistController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateWhitelistDto"}}}},"responses":{"201":{"description":""}},"tags":["WhiteList"]},"get":{"operationId":"WhitelistController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["WhiteList"]}},"/api/whitelist/{id}":{"get":{"operationId":"WhitelistController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["WhiteList"]},"patch":{"operationId":"WhitelistController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateWhitelistDto"}}}},"responses":{"200":{"description":""}},"tags":["WhiteList"]},"delete":{"operationId":"WhitelistController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["WhiteList"]}},"/api/project":{"get":{"operationId":"ProjectController_getData","parameters":[],"responses":{"200":{"description":"test get Project","schema":{"$ref":"#/components/schemas/Project"},"content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Project"}}}}}},"tags":["Project"]},"post":{"operationId":"ProjectController_createData","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAppDto"}}}},"responses":{"200":{"description":"test create Project","schema":{"$ref":"#/components/schemas/Project"},"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Project"}}}}},"tags":["Project"],"security":[{"bearer":[]}]},"delete":{"operationId":"ProjectController_removeData","parameters":[],"responses":{"200":{"description":"test remove app","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessResponse"}}}}},"tags":["Project"],"security":[{"bearer":[]}]}},"/api/project/{id}":{"get":{"operationId":"ProjectController_getDataById","parameters":[{"name":"id","required":true,"in":"path","schema":{}}],"responses":{"200":{"description":"test get Project","schema":{"$ref":"#/components/schemas/Project"},"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Project"}}}}},"tags":["Project"]}}},"info":{"title":"API","description":"API description","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"Bearer","bearerFormat":"Token","name":"Authorization","type":"http"}},"schemas":{"CreateWhitelistDto":{"type":"object","properties":{"name":{"type":"string"}},"required":["name"]},"UpdateWhitelistDto":{"type":"object","properties":{}},"Project":{"type":"object","properties":{"name":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"}},"required":["name","createdAt","updatedAt"]},"CreateAppDto":{"type":"object","properties":{"name":{"type":"string"}},"required":["name"]},"SuccessResponse":{"type":"object","properties":{"message":{"type":"string","example":"Success!!"}},"required":["message"]}}}}